Added authentication for controllers

+++ app/controllers/routines_controller.rb
@@ -13,38 +13,55 @@ class RoutinesController < ApplicationController
     end 
 
     post '/routines' do 
-        @routine = Routine.new 
-        @routine.name = params[:name] 
-        @routine.why = params[:why] 
-        @routine.duration = params[:duration]
-        @routine.user_id = current_user.id
-        @routine.save
-        
-        redirect '/routines'
+        @routine = Routine.new(name: params[:name], why: params[:why], duration: params[:duration], user_id: session[:user_id])
+    
+        if @routine.save
+            flash[:message] = "Nice! New routine! Let's add some habits!"
+            redirect "/routines/#{@routine.id}"
+        else 
+            flash[:error] = "Cannot create routine: #{@routine.errors.full_messages.to_sentence}"
+            redirect '/routines/new' 
+        end 
     end 
 
 
     get '/routines/:id' do 
         @user = current_user 
         @routine = Routine.find_by_id(params[:id])
-    
-        erb :'routines/show.html'
+        if can_edit_routine(@routine) 
+            session[:routine_id] = @routine[:id] 
+            erb :'routines/show.html'
+        else 
+            flash[:error] = "cannot view another users routine" 
+            redirect '/routines' 
+        end 
     end 
 
     get '/routines/:id/edit' do 
         @routine = Routine.find_by_id(params[:id])
-        erb :'routines/edit.html'
+        if can_edit_routine(@routine)
+            erb :'routines/edit.html'
+        else 
+            flash[:error] = "Cannot edit this Routine, it is not yours!"
+            redirect "/routines"
+        end 
+
     end 
 
     patch '/routines/:id' do 
-        @routine = Routine.find_by_id(params[:id]) 
-        @routine.update(params[:routine])
+        @routine = Routine.find_by_id(params[:id])
+        @routine.update(params[:routine]) 
+        redirect '/routines'
     end 
 
     delete '/routines/:id' do 
         routine = Routine.find_by_id(params[:id])
-        routine.destroy 
-
-        redirect '/routines'
+        if can_edit_routine(routine) 
+            routine.destroy 
+            redirect '/routines'
+        else 
+            flash[:error] = "cannot delete another users routine"
+            redirect '/routines' 
+        end 
     end 
 end
